+++ Cargo.toml
@@ -28,6 +28,7 @@ solana-cli-config = "^2.1"
 solana-program = "^2.1"
 solana-sdk = "^2.1"
 spl-token = { features = ["no-entrypoint"], version = "^4" }
+spl-token-2022 = { features = ["no-entrypoint"], version = "^7" }
 spl-associated-token-account = { features = [ "no-entrypoint" ], version = "6.0" }
 thiserror = "1.0.57"
 tokio = "1.35"
