+++ lib/Cargo.toml
@@ -11,14 +11,15 @@ repository.workspace = true
 keywords.workspace = true
 
 [features]
-deafult = []
-spl = ["spl-token", "spl-associated-token-account"]
+default = ["spl"]
+spl = ["spl-token", "spl-token-2022", "spl-associated-token-account"]
 
 [dependencies]
 bytemuck.workspace = true
 num_enum.workspace = true
 solana-program.workspace = true
 spl-token = { workspace = true, optional = true }
+spl-token-2022 = { workspace = true, optional = true }
 spl-associated-token-account = { workspace = true, optional = true }
 thiserror.workspace = true
