+++ lib/src/account/mod.rs
@@ -1,0 +1,6 @@
+mod cpi;
+mod traits;
+mod validation;
+
+pub use cpi::*;
+pub use traits::*;
