+++ lib/src/spl/cpi.rs
@@ -1,233 +1,7 @@
-use bytemuck::Pod;
-use solana_program::{
-    account_info::AccountInfo, entrypoint::ProgramResult, instruction::Instruction, pubkey::Pubkey,
-    rent::Rent, sysvar::Sysvar,
-};
+use solana_program::{account_info::AccountInfo, entrypoint::ProgramResult, pubkey::Pubkey};
 
-use crate::Discriminator;
+use crate::account::invoke_signed_with_bump;
 
-/// Creates a new account.
-#[inline(always)]
-pub fn create_account<'a, 'info>(
-    from_pubkey: &'a AccountInfo<'info>,
-    to_pubkey: &'a AccountInfo<'info>,
-    system_program: &'a AccountInfo<'info>,
-    space: usize,
-    owner: &Pubkey,
-) -> ProgramResult {
-    let lamports_required = (Rent::get()?).minimum_balance(space);
-
-    solana_program::program::invoke(
-        &solana_program::system_instruction::create_account(
-            from_pubkey.key,
-            to_pubkey.key,
-            lamports_required,
-            space as u64,
-            owner,
-        ),
-        &[
-            from_pubkey.clone(),
-            to_pubkey.clone(),
-            system_program.clone(),
-        ],
-    )?;
-
-    Ok(())
-}
-
-/// Creates a new program account.
-#[inline(always)]
-pub fn create_program_account<'a, 'info, T: Discriminator + Pod>(
-    target_account: &'a AccountInfo<'info>,
-    system_program: &'a AccountInfo<'info>,
-    payer: &'a AccountInfo<'info>,
-    owner: &Pubkey,
-    seeds: &[&[u8]],
-) -> ProgramResult {
-    create_program_account_with_bump::<T>(
-        target_account,
-        system_program,
-        payer,
-        owner,
-        seeds,
-        Pubkey::find_program_address(seeds, owner).1,
-    )
-}
-
-/// Creates a new program account with user-provided bump.
-#[inline(always)]
-pub fn create_program_account_with_bump<'a, 'info, T: Discriminator + Pod>(
-    target_account: &'a AccountInfo<'info>,
-    system_program: &'a AccountInfo<'info>,
-    payer: &'a AccountInfo<'info>,
-    owner: &Pubkey,
-    seeds: &[&[u8]],
-    bump: u8,
-) -> ProgramResult {
-    // Allocate space.
-    allocate_account_with_bump(
-        target_account,
-        system_program,
-        payer,
-        8 + std::mem::size_of::<T>(),
-        owner,
-        seeds,
-        bump,
-    )?;
-
-    // Set discriminator.
-    let mut data = target_account.data.borrow_mut();
-    data[0] = T::discriminator();
-
-    Ok(())
-}
-
-/// Allocates space for a new program account.
-#[inline(always)]
-pub fn allocate_account<'a, 'info>(
-    target_account: &'a AccountInfo<'info>,
-    system_program: &'a AccountInfo<'info>,
-    payer: &'a AccountInfo<'info>,
-    space: usize,
-    owner: &Pubkey,
-    seeds: &[&[u8]],
-) -> ProgramResult {
-    allocate_account_with_bump(
-        target_account,
-        system_program,
-        payer,
-        space,
-        owner,
-        seeds,
-        Pubkey::find_program_address(seeds, owner).1,
-    )
-}
-
-/// Allocates space for a new program account with user-provided bump.
-#[inline(always)]
-pub fn allocate_account_with_bump<'a, 'info>(
-    target_account: &'a AccountInfo<'info>,
-    system_program: &'a AccountInfo<'info>,
-    payer: &'a AccountInfo<'info>,
-    space: usize,
-    owner: &Pubkey,
-    seeds: &[&[u8]],
-    bump: u8,
-) -> ProgramResult {
-    // Combine seeds
-    let bump: &[u8] = &[bump];
-    let mut combined_seeds = Vec::with_capacity(seeds.len() + 1);
-    combined_seeds.extend_from_slice(seeds);
-    combined_seeds.push(bump);
-    let seeds = combined_seeds.as_slice();
-
-    // Allocate space for account
-    let rent = Rent::get()?;
-    if target_account.lamports().eq(&0) {
-        // If balance is zero, create account
-        solana_program::program::invoke_signed(
-            &solana_program::system_instruction::create_account(
-                payer.key,
-                target_account.key,
-                rent.minimum_balance(space),
-                space as u64,
-                owner,
-            ),
-            &[
-                payer.clone(),
-                target_account.clone(),
-                system_program.clone(),
-            ],
-            &[seeds],
-        )?;
-    } else {
-        // Otherwise, if balance is nonzero:
-
-        // 1) transfer sufficient lamports for rent exemption
-        let rent_exempt_balance = rent
-            .minimum_balance(space)
-            .saturating_sub(target_account.lamports());
-        if rent_exempt_balance.gt(&0) {
-            solana_program::program::invoke(
-                &solana_program::system_instruction::transfer(
-                    payer.key,
-                    target_account.key,
-                    rent_exempt_balance,
-                ),
-                &[
-                    payer.clone(),
-                    target_account.clone(),
-                    system_program.clone(),
-                ],
-            )?;
-        }
-
-        // 2) allocate space for the account
-        solana_program::program::invoke_signed(
-            &solana_program::system_instruction::allocate(target_account.key, space as u64),
-            &[target_account.clone(), system_program.clone()],
-            &[seeds],
-        )?;
-
-        // 3) assign our program as the owner
-        solana_program::program::invoke_signed(
-            &solana_program::system_instruction::assign(target_account.key, owner),
-            &[target_account.clone(), system_program.clone()],
-            &[seeds],
-        )?;
-    }
-
-    Ok(())
-}
-
-/// Closes an account and returns the remaining rent lamports to the provided recipient.
-#[inline(always)]
-pub fn close_account<'info>(
-    account_info: &AccountInfo<'info>,
-    recipient: &AccountInfo<'info>,
-) -> ProgramResult {
-    // Realloc data to zero.
-    account_info.realloc(0, true)?;
-
-    // Return rent lamports.
-    **recipient.lamports.borrow_mut() += account_info.lamports();
-    **account_info.lamports.borrow_mut() = 0;
-
-    Ok(())
-}
-
-/// Invokes a CPI with provided signer seeds and program id.
-#[inline(always)]
-pub fn invoke_signed<'info>(
-    instruction: &Instruction,
-    account_infos: &[AccountInfo<'info>],
-    program_id: &Pubkey,
-    seeds: &[&[u8]],
-) -> ProgramResult {
-    let bump = Pubkey::find_program_address(seeds, program_id).1;
-    invoke_signed_with_bump(instruction, account_infos, seeds, bump)
-}
-
-/// Invokes a CPI with the provided signer seeds and bump.
-#[inline(always)]
-pub fn invoke_signed_with_bump<'info>(
-    instruction: &Instruction,
-    account_infos: &[AccountInfo<'info>],
-    seeds: &[&[u8]],
-    bump: u8,
-) -> ProgramResult {
-    // Combine seeds
-    let bump: &[u8] = &[bump];
-    let mut combined_seeds = Vec::with_capacity(seeds.len() + 1);
-    combined_seeds.extend_from_slice(seeds);
-    combined_seeds.push(bump);
-    let seeds = combined_seeds.as_slice();
-
-    // Invoke CPI
-    solana_program::program::invoke_signed(instruction, account_infos, &[seeds])
-}
-
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn create_associated_token_account<'info>(
     funder_info: &AccountInfo<'info>,
@@ -257,7 +31,6 @@ pub fn create_associated_token_account<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn close_token_account<'info>(
     account_info: &AccountInfo<'info>,
@@ -282,7 +55,6 @@ pub fn close_token_account<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn close_token_account_signed<'info>(
     account_info: &AccountInfo<'info>,
@@ -302,7 +74,6 @@ pub fn close_token_account_signed<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn close_token_account_signed_with_bump<'info>(
     account_info: &AccountInfo<'info>,
@@ -331,7 +102,6 @@ pub fn close_token_account_signed_with_bump<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn transfer<'info>(
     authority_info: &AccountInfo<'info>,
@@ -358,7 +128,6 @@ pub fn transfer<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn transfer_signed<'info>(
     authority_info: &AccountInfo<'info>,
@@ -380,7 +149,6 @@ pub fn transfer_signed<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn transfer_signed_with_bump<'info>(
     authority_info: &AccountInfo<'info>,
@@ -411,7 +179,6 @@ pub fn transfer_signed_with_bump<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn transfer_checked<'info>(
     authority_info: &AccountInfo<'info>,
@@ -423,8 +190,8 @@ pub fn transfer_checked<'info>(
     decimals: u8,
 ) -> ProgramResult {
     solana_program::program::invoke(
-        &spl_token::instruction::transfer_checked(
-            &spl_token::ID,
+        &spl_token_2022::instruction::transfer_checked(
+            &token_program.key,
             from_info.key,
             mint_info.key,
             to_info.key,
@@ -443,7 +210,6 @@ pub fn transfer_checked<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn transfer_checked_signed<'info>(
     authority_info: &AccountInfo<'info>,
@@ -469,7 +235,6 @@ pub fn transfer_checked_signed<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn transfer_checked_signed_with_bump<'info>(
     authority_info: &AccountInfo<'info>,
@@ -483,8 +248,8 @@ pub fn transfer_checked_signed_with_bump<'info>(
     bump: u8,
 ) -> ProgramResult {
     invoke_signed_with_bump(
-        &spl_token::instruction::transfer_checked(
-            &spl_token::ID,
+        &spl_token_2022::instruction::transfer_checked(
+            &token_program.key,
             from_info.key,
             mint_info.key,
             to_info.key,
@@ -505,7 +270,6 @@ pub fn transfer_checked_signed_with_bump<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn mint_to_signed<'info>(
     mint_info: &AccountInfo<'info>,
@@ -527,7 +291,6 @@ pub fn mint_to_signed<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn mint_to_signed_with_bump<'info>(
     mint_info: &AccountInfo<'info>,
@@ -539,8 +302,8 @@ pub fn mint_to_signed_with_bump<'info>(
     bump: u8,
 ) -> ProgramResult {
     invoke_signed_with_bump(
-        &spl_token::instruction::mint_to(
-            &spl_token::ID,
+        &spl_token_2022::instruction::mint_to(
+            &token_program.key,
             mint_info.key,
             to_info.key,
             authority_info.key,
@@ -558,7 +321,6 @@ pub fn mint_to_signed_with_bump<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn mint_to_checked_signed<'info>(
     mint_info: &AccountInfo<'info>,
@@ -582,7 +344,6 @@ pub fn mint_to_checked_signed<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn mint_to_checked_signed_with_bump<'info>(
     mint_info: &AccountInfo<'info>,
@@ -595,8 +356,8 @@ pub fn mint_to_checked_signed_with_bump<'info>(
     bump: u8,
 ) -> ProgramResult {
     invoke_signed_with_bump(
-        &spl_token::instruction::mint_to_checked(
-            &spl_token::ID,
+        &spl_token_2022::instruction::mint_to_checked(
+            &token_program.key,
             mint_info.key,
             to_info.key,
             authority_info.key,
@@ -615,7 +376,6 @@ pub fn mint_to_checked_signed_with_bump<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn burn<'info>(
     token_account_info: &AccountInfo<'info>,
@@ -625,8 +385,8 @@ pub fn burn<'info>(
     amount: u64,
 ) -> ProgramResult {
     solana_program::program::invoke(
-        &spl_token::instruction::burn(
-            &spl_token::ID,
+        &spl_token_2022::instruction::burn(
+            &token_program.key,
             token_account_info.key,
             mint_info.key,
             authority_info.key,
@@ -642,7 +402,6 @@ pub fn burn<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn burn_signed<'info>(
     token_account_info: &AccountInfo<'info>,
@@ -664,7 +423,6 @@ pub fn burn_signed<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn burn_signed_with_bump<'info>(
     token_account_info: &AccountInfo<'info>,
@@ -676,8 +434,8 @@ pub fn burn_signed_with_bump<'info>(
     bump: u8,
 ) -> ProgramResult {
     invoke_signed_with_bump(
-        &spl_token::instruction::burn(
-            &spl_token::ID,
+        &spl_token_2022::instruction::burn(
+            &token_program.key,
             token_account_info.key,
             mint_info.key,
             authority_info.key,
@@ -695,7 +453,6 @@ pub fn burn_signed_with_bump<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn burn_checked<'info>(
     token_account_info: &AccountInfo<'info>,
@@ -706,8 +463,8 @@ pub fn burn_checked<'info>(
     decimals: u8,
 ) -> ProgramResult {
     solana_program::program::invoke(
-        &spl_token::instruction::burn_checked(
-            &spl_token::ID,
+        &spl_token_2022::instruction::burn_checked(
+            &token_program.key,
             token_account_info.key,
             mint_info.key,
             authority_info.key,
@@ -724,7 +481,6 @@ pub fn burn_checked<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn burn_checked_signed<'info>(
     token_account_info: &AccountInfo<'info>,
@@ -748,7 +504,6 @@ pub fn burn_checked_signed<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn burn_checked_signed_with_bump<'info>(
     token_account_info: &AccountInfo<'info>,
@@ -761,8 +516,8 @@ pub fn burn_checked_signed_with_bump<'info>(
     bump: u8,
 ) -> ProgramResult {
     invoke_signed_with_bump(
-        &spl_token::instruction::burn_checked(
-            &spl_token::ID,
+        &spl_token_2022::instruction::burn_checked(
+            &token_program.key,
             token_account_info.key,
             mint_info.key,
             authority_info.key,
@@ -781,7 +536,6 @@ pub fn burn_checked_signed_with_bump<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn freeze<'info>(
     account_info: &AccountInfo<'info>,
@@ -791,8 +545,8 @@ pub fn freeze<'info>(
     token_program: &AccountInfo<'info>,
 ) -> ProgramResult {
     solana_program::program::invoke(
-        &spl_token::instruction::freeze_account(
-            &spl_token::ID,
+        &spl_token_2022::instruction::freeze_account(
+            &token_program.key,
             account_info.key,
             mint_info.key,
             owner_info.key,
@@ -808,7 +562,6 @@ pub fn freeze<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn freeze_signed<'info>(
     account_info: &AccountInfo<'info>,
@@ -830,7 +583,6 @@ pub fn freeze_signed<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn freeze_signed_with_bump<'info>(
     account_info: &AccountInfo<'info>,
@@ -842,8 +594,8 @@ pub fn freeze_signed_with_bump<'info>(
     bump: u8,
 ) -> ProgramResult {
     invoke_signed_with_bump(
-        &spl_token::instruction::freeze_account(
-            &spl_token::ID,
+        &spl_token_2022::instruction::freeze_account(
+            &token_program.key,
             account_info.key,
             mint_info.key,
             owner_info.key,
@@ -861,7 +613,6 @@ pub fn freeze_signed_with_bump<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn initialize_mint<'info>(
     mint_info: &AccountInfo<'info>,
@@ -872,8 +623,8 @@ pub fn initialize_mint<'info>(
     decimals: u8,
 ) -> ProgramResult {
     solana_program::program::invoke(
-        &spl_token::instruction::initialize_mint(
-            &spl_token::ID,
+        &spl_token_2022::instruction::initialize_mint(
+            &token_program.key,
             mint_info.key,
             mint_authority_info.key,
             freeze_authority_info.map(|i| i.key),
@@ -888,7 +639,6 @@ pub fn initialize_mint<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn initialize_mint_signed<'info>(
     mint_info: &AccountInfo<'info>,
@@ -912,7 +662,6 @@ pub fn initialize_mint_signed<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn initialize_mint_signed_with_bump<'info>(
     mint_info: &AccountInfo<'info>,
@@ -925,8 +674,8 @@ pub fn initialize_mint_signed_with_bump<'info>(
     bump: u8,
 ) -> ProgramResult {
     invoke_signed_with_bump(
-        &spl_token::instruction::initialize_mint(
-            &spl_token::ID,
+        &spl_token_2022::instruction::initialize_mint(
+            &token_program.key,
             mint_info.key,
             mint_authority_info.key,
             freeze_authority_info.map(|i| i.key),
@@ -945,7 +694,6 @@ pub fn initialize_mint_signed_with_bump<'info>(
 
 /// Thaws a frozen SPL token account
 ///
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn thaw_account<'info>(
     token_account_info: &AccountInfo<'info>,
@@ -954,8 +702,8 @@ pub fn thaw_account<'info>(
     token_program: &AccountInfo<'info>,
 ) -> ProgramResult {
     solana_program::program::invoke(
-        &spl_token::instruction::thaw_account(
-            &spl_token::id(),
+        &spl_token_2022::instruction::thaw_account(
+            &token_program.key,
             token_account_info.key,
             mint_info.key,
             authority_info.key,
@@ -972,7 +720,6 @@ pub fn thaw_account<'info>(
 
 /// Thaws a frozen SPL token account using signed account
 ///
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn thaw_account_signed<'info>(
     token_account_info: &AccountInfo<'info>,
@@ -994,7 +741,6 @@ pub fn thaw_account_signed<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn thaw_account_signed_with_bump<'info>(
     token_account_info: &AccountInfo<'info>,
@@ -1006,8 +752,8 @@ pub fn thaw_account_signed_with_bump<'info>(
     bump: u8,
 ) -> ProgramResult {
     invoke_signed_with_bump(
-        &spl_token::instruction::thaw_account(
-            &spl_token::id(),
+        &spl_token_2022::instruction::thaw_account(
+            &token_program.key,
             token_account_info.key,
             mint_info.key,
             owner_info.key,
@@ -1027,18 +773,17 @@ pub fn thaw_account_signed_with_bump<'info>(
 
 /// Set authority for an SPL token mint
 ///
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn set_authority<'info>(
     account_or_mint: &AccountInfo<'info>,
     authority_info: &AccountInfo<'info>,
     new_authority_info: Option<&AccountInfo<'info>>,
-    authority_type: spl_token::instruction::AuthorityType,
+    authority_type: spl_token_2022::instruction::AuthorityType,
     token_program: &AccountInfo<'info>,
 ) -> ProgramResult {
     solana_program::program::invoke(
-        &spl_token::instruction::set_authority(
-            &spl_token::id(),
+        &spl_token_2022::instruction::set_authority(
+            &token_program.key,
             account_or_mint.key,
             new_authority_info.map(|i| i.key),
             authority_type,
@@ -1055,13 +800,12 @@ pub fn set_authority<'info>(
 
 /// Set authority using signer seeds
 ///
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn set_authority_signed<'info>(
     account_or_mint: &AccountInfo<'info>,
     authority_info: &AccountInfo<'info>,
     new_authority_info: Option<&AccountInfo<'info>>,
-    authority_type: spl_token::instruction::AuthorityType,
+    authority_type: spl_token_2022::instruction::AuthorityType,
     token_program: &AccountInfo<'info>,
     seeds: &[&[u8]],
 ) -> ProgramResult {
@@ -1077,20 +821,19 @@ pub fn set_authority_signed<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn set_authority_signed_with_bump<'info>(
     account_or_mint: &AccountInfo<'info>,
     authority_info: &AccountInfo<'info>,
     new_authority_info: Option<&AccountInfo<'info>>,
-    authority_type: spl_token::instruction::AuthorityType,
+    authority_type: spl_token_2022::instruction::AuthorityType,
     token_program: &AccountInfo<'info>,
     seeds: &[&[u8]],
     bump: u8,
 ) -> ProgramResult {
     invoke_signed_with_bump(
-        &spl_token::instruction::set_authority(
-            &spl_token::id(),
+        &spl_token_2022::instruction::set_authority(
+            &token_program.key,
             account_or_mint.key,
             new_authority_info.map(|i| i.key),
             authority_type,
@@ -1109,7 +852,6 @@ pub fn set_authority_signed_with_bump<'info>(
 
 /// Revoke
 ///
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn revoke<'info>(
     source_info: &AccountInfo<'info>,
@@ -1117,8 +859,8 @@ pub fn revoke<'info>(
     token_program: &AccountInfo<'info>,
 ) -> ProgramResult {
     solana_program::program::invoke(
-        &spl_token::instruction::revoke(
-            &spl_token::id(),
+        &spl_token_2022::instruction::revoke(
+            &token_program.key,
             source_info.key,
             authority_info.key,
             &[authority_info.key],
@@ -1133,7 +875,6 @@ pub fn revoke<'info>(
 
 /// Revoke with signer seeds
 ///
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn revoke_signed<'info>(
     source_info: &AccountInfo<'info>,
@@ -1145,7 +886,6 @@ pub fn revoke_signed<'info>(
     revoke_signed_with_bump(source_info, authority_info, token_program, seeds, bump)
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn revoke_signed_with_bump<'info>(
     source_info: &AccountInfo<'info>,
@@ -1155,8 +895,8 @@ pub fn revoke_signed_with_bump<'info>(
     bump: u8,
 ) -> ProgramResult {
     invoke_signed_with_bump(
-        &spl_token::instruction::revoke(
-            &spl_token::id(),
+        &spl_token_2022::instruction::revoke(
+            &token_program.key,
             source_info.key,
             authority_info.key,
             &[authority_info.key],
@@ -1173,7 +913,6 @@ pub fn revoke_signed_with_bump<'info>(
 
 ///  /// Approves a delegate.  A delegate is given the authority over tokens on
 /// behalf of the source account's owner.
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn approve<'info>(
     source_info: &AccountInfo<'info>,
@@ -1202,7 +941,6 @@ pub fn approve<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn approve_signed<'info>(
     source_info: &AccountInfo<'info>,
@@ -1222,7 +960,6 @@ pub fn approve_signed<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn approve_signed_with_bump<'info>(
     source_info: &AccountInfo<'info>,
@@ -1233,8 +970,8 @@ pub fn approve_signed_with_bump<'info>(
     bump: u8,
 ) -> ProgramResult {
     invoke_signed_with_bump(
-        &spl_token::instruction::approve(
-            &spl_token::ID,
+        &spl_token_2022::instruction::approve(
+            &token_program.key,
             source_info.key,
             delegate_info.key,
             delegate_info.key,
@@ -1251,7 +988,6 @@ pub fn approve_signed_with_bump<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn approve_checked<'info>(
     source_info: &AccountInfo<'info>,
@@ -1262,8 +998,8 @@ pub fn approve_checked<'info>(
     decimals: u8,
 ) -> ProgramResult {
     solana_program::program::invoke(
-        &spl_token::instruction::approve_checked(
-            &spl_token::ID,
+        &spl_token_2022::instruction::approve_checked(
+            &token_program.key,
             mint_info.key,
             source_info.key,
             delegate_info.key,
@@ -1281,7 +1017,6 @@ pub fn approve_checked<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn approve_checked_signed<'info>(
     source_info: &AccountInfo<'info>,
@@ -1301,20 +1036,20 @@ pub fn approve_checked_signed<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn approve_checked_signed_with_bump<'info>(
     source_info: &AccountInfo<'info>,
     mint_info: &AccountInfo<'info>,
     delegate_info: &AccountInfo<'info>,
+    token_program: &AccountInfo<'info>,
     decimals: u8,
     amount: u64,
     seeds: &[&[u8]],
     bump: u8,
 ) -> ProgramResult {
     invoke_signed_with_bump(
-        &spl_token::instruction::approve_checked(
-            &spl_token::ID,
+        &spl_token_2022::instruction::approve_checked(
+            &token_program.key,
             source_info.key,
             mint_info.key,
             delegate_info.key,
@@ -1330,7 +1065,6 @@ pub fn approve_checked_signed_with_bump<'info>(
 }
 
 /// Initializes a multisignature account with N provided signers.
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn initialize_multisig<'info>(
     multisig_info: &AccountInfo<'info>,
@@ -1339,8 +1073,8 @@ pub fn initialize_multisig<'info>(
     n: u8,
 ) -> ProgramResult {
     solana_program::program::invoke(
-        &spl_token::instruction::initialize_multisig(
-            &spl_token::ID,
+        &spl_token_2022::instruction::initialize_multisig(
+            &token_program.key,
             multisig_info.key,
             &[signer_info.key],
             n,
@@ -1349,7 +1083,6 @@ pub fn initialize_multisig<'info>(
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn initialize_multisig_signed<'info>(
     multisig_info: &AccountInfo<'info>,
@@ -1362,7 +1095,6 @@ pub fn initialize_multisig_signed<'info>(
     initialize_multisig_signed_with_bump(multisig_info, signer_info, token_program, m, seeds, bump)
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn initialize_multisig_signed_with_bump<'info>(
     multisig_info: &AccountInfo<'info>,
@@ -1373,8 +1105,8 @@ pub fn initialize_multisig_signed_with_bump<'info>(
     bump: u8,
 ) -> ProgramResult {
     invoke_signed_with_bump(
-        &spl_token::instruction::initialize_multisig(
-            &spl_token::ID,
+        &spl_token_2022::instruction::initialize_multisig(
+            &token_program.key,
             multisig_info.key,
             &[signer_info.key],
             m,
@@ -1386,22 +1118,20 @@ pub fn initialize_multisig_signed_with_bump<'info>(
 }
 
 ///Like InitializeAccount2, but does not require the Rent sysvar to be provided
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn sync_native(account_info: &AccountInfo) -> ProgramResult {
     solana_program::program::invoke(
-        &spl_token::instruction::sync_native(&spl_token::ID, account_info.key)?,
+        &spl_token_2022::instruction::sync_native(account_info.owner, account_info.key)?,
         &[account_info.clone()],
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
 pub fn sync_native_signed(account_info: &AccountInfo, seeds: &[&[u8]]) -> ProgramResult {
     let bump = Pubkey::find_program_address(seeds, &spl_token::ID).1;
     sync_native_signed_with_bump(account_info, seeds, bump)
 }
-#[cfg(feature = "spl")]
+
 #[inline(always)]
 pub fn sync_native_signed_with_bump(
     account_info: &AccountInfo,
@@ -1409,18 +1139,24 @@ pub fn sync_native_signed_with_bump(
     bump: u8,
 ) -> ProgramResult {
     invoke_signed_with_bump(
-        &spl_token::instruction::sync_native(&spl_token::ID, account_info.key)?,
+        &spl_token_2022::instruction::sync_native(account_info.owner, account_info.key)?,
         &[account_info.clone()],
         seeds,
         bump,
     )
 }
 
-#[cfg(feature = "spl")]
 #[inline(always)]
-pub fn get_account_data_size(mint_info: &AccountInfo) -> ProgramResult {
+pub fn get_account_data_size(
+    mint_info: &AccountInfo,
+    extension_types: &[spl_token_2022::extension::ExtensionType],
+) -> ProgramResult {
     solana_program::program::invoke(
-        &spl_token::instruction::get_account_data_size(&spl_token::ID, mint_info.key)?,
+        &spl_token_2022::instruction::get_account_data_size(
+            mint_info.owner,
+            mint_info.key,
+            extension_types,
+        )?,
         &[mint_info.clone()],
     )
 }
